연락처 확인 MainActivity.kt 쪽으로 이동 및 소규모 수정
